dns: move OpenRawSock in a separate pkg
It will be used by dns and snisnoop gadgets.
